[TD-5505]<fix>: windows compilation error

+++ src/os/src/detail/osTime.c
@@ -97,7 +97,7 @@ int32_t taosGetTimestampSec() { return (int32_t)time(NULL); }
 int32_t taosParseTime(char* timestr, int64_t* time, int32_t len, int32_t timePrec, int8_t day_light) {
   /* parse datatime string in with tz */
   char *seg = forwardToTimeStringEnd(timestr);
-  int32_t seg_len = len - (seg - timestr);
+  int32_t seg_len = len - (int32_t)(seg - timestr);
   if (strnchr(timestr, 'T', len, false) != NULL) {
     return parseTimeWithTz(timestr, time, timePrec, 'T');
   } else if (strnchr(timestr, ' ', len, false) != NULL &&
